+++ met-web/tests/unit/components/engagement/EngagementForm.test.tsx
@@ -1,8 +1,8 @@
 import { render, waitFor, screen, fireEvent } from '@testing-library/react';
 import React from 'react';
 import '@testing-library/jest-dom';
-import EngagementForm from '../../../src/components/engagement/form';
-import { setupEnv } from './setEnvVars';
+import EngagementForm from '../../../../src/components/engagement/form';
+import { setupEnv } from '../setEnvVars';
 import * as reactRedux from 'react-redux';
 import * as reactRouter from 'react-router';
 import * as engagementService from 'services/engagementService';
@@ -10,10 +10,9 @@ import * as widgetService from 'services/widgetService';
 import * as notificationSlice from 'services/notificationService/notificationSlice';
 import * as notificationModalSlice from 'services/notificationModalService/notificationModalSlice';
 import { createDefaultSurvey, Survey } from 'models/survey';
-import { Widget, WidgetItem, WidgetType } from 'models/widget';
-import { Contact } from 'models/contact';
+import { WidgetType } from 'models/widget';
 import { Box } from '@mui/material';
-import { engagement } from './factory';
+import { engagement } from '../factory';
 
 const survey: Survey = {
     ...createDefaultSurvey(),
@@ -23,46 +22,6 @@ const survey: Survey = {
 };
 
 const surveys = [survey];
-
-const mockContact: Contact = {
-    id: 1,
-    name: 'Jace',
-    title: 'prince',
-    phone_number: '123-123-1234',
-    email: 'jace@gmail.com',
-    address: 'Dragonstone, Westeros',
-    bio: 'Jacaerys Targaryen is the son and heir of Rhaenyra Targaryen',
-    avatar_filename: '',
-    avatar_url: '',
-};
-
-const contactWidgetItem: WidgetItem = {
-    id: 1,
-    widget_id: 1,
-    widget_data_id: 1,
-    sort_index: 1,
-};
-
-const whoIsListeningWidget: Widget = {
-    id: 1,
-    widget_type_id: WidgetType.WhoIsListening,
-    engagement_id: 1,
-    items: [contactWidgetItem],
-};
-
-const phaseWidgetItem: WidgetItem = {
-    id: 2,
-    widget_id: 2,
-    widget_data_id: 0,
-    sort_index: 1,
-};
-
-const phasesWidget: Widget = {
-    id: 2,
-    widget_type_id: WidgetType.Phases,
-    engagement_id: 1,
-    items: [phaseWidgetItem],
-};
 
 jest.mock('react-dnd', () => ({
     ...jest.requireActual('react-dnd'),
@@ -104,7 +63,7 @@ jest.mock('apiManager/apiSlices/contacts', () => ({
     useLazyGetContactQuery: () => [...mockLazyGetContactQuery()],
 }));
 
-const mockWidgetsRtkUnwrap = jest.fn(() => Promise.resolve([whoIsListeningWidget, phasesWidget]));
+const mockWidgetsRtkUnwrap = jest.fn(() => Promise.resolve([]));
 const mockWidgetsRtkTrigger = () => {
     return {
         unwrap: mockWidgetsRtkUnwrap,
@@ -136,12 +95,9 @@ describe('Engagement form page tests', () => {
     const postEngagementMock = jest
         .spyOn(engagementService, 'postEngagement')
         .mockReturnValue(Promise.resolve(engagement));
-    const postWidgetMock = jest
-        .spyOn(widgetService, 'postWidget')
-        .mockReturnValue(Promise.resolve(whoIsListeningWidget));
-    const postWidgetItemMock = jest
-        .spyOn(widgetService, 'postWidgetItem')
-        .mockReturnValue(Promise.resolve(contactWidgetItem));
+    const getWidgetsMock = jest
+        .spyOn(widgetService, 'getWidgets')
+        .mockReturnValue(Promise.resolve([]));
 
     beforeEach(() => {
         setupEnv();
@@ -363,145 +319,6 @@ describe('Engagement form page tests', () => {
             expect(screen.getByText('Select Widget')).toBeVisible();
             expect(screen.getByTestId(`widget-drawer-option/${WidgetType.WhoIsListening}`));
             expect(screen.getByTestId(`widget-drawer-option/${WidgetType.Phases}`));
-        });
-    });
-
-    test('Who is listening widget is created when option is clicked', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        getEngagementMock.mockReturnValueOnce(
-            Promise.resolve({
-                ...engagement,
-                surveys: surveys,
-            }),
-        );
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([]));
-        postWidgetMock.mockReturnValue(Promise.resolve(whoIsListeningWidget));
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([whoIsListeningWidget]));
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-
-        const addWidgetButton = screen.getByText('Add Widget');
-        fireEvent.click(addWidgetButton);
-
-        await waitFor(() => {
-            expect(screen.getByText('Select Widget')).toBeVisible();
-        });
-
-        const whoIsListeningOption = screen.getByTestId(`widget-drawer-option/${WidgetType.WhoIsListening}`);
-        fireEvent.click(whoIsListeningOption);
-
-        await waitFor(() => {
-            expect(screen.getByText('Add This Contact')).toBeVisible();
-        });
-        expect(postWidgetMock).toHaveBeenNthCalledWith(1, engagement.id, {
-            widget_type_id: WidgetType.WhoIsListening,
-            engagement_id: engagement.id,
-        });
-        expect(mockWidgetsRtkUnwrap).toHaveBeenCalledTimes(2);
-        expect(screen.getByText('Add This Contact')).toBeVisible();
-        expect(screen.getByText('Select Existing Contact')).toBeVisible();
-    });
-
-    test('Add contact drawer appears', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        getEngagementMock.mockReturnValueOnce(
-            Promise.resolve({
-                ...engagement,
-                surveys: surveys,
-            }),
-        );
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([]));
-        postWidgetMock.mockReturnValue(Promise.resolve(whoIsListeningWidget));
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([whoIsListeningWidget]));
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-
-        const addWidgetButton = screen.getByText('Add Widget');
-        fireEvent.click(addWidgetButton);
-
-        await waitFor(() => {
-            expect(screen.getByText('Select Widget')).toBeVisible();
-        });
-
-        const whoIsListeningOption = screen.getByTestId(`widget-drawer-option/${WidgetType.WhoIsListening}`);
-        fireEvent.click(whoIsListeningOption);
-
-        await waitFor(() => {
-            expect(screen.getByText('Create New Contact')).toBeVisible();
-            expect(screen.getByText(mockContact.email)).toBeVisible();
-        });
-
-        expect(mockContactsRtkUnwrap).toHaveBeenCalled();
-
-        const createContactButton = screen.getByText('Create New Contact');
-        fireEvent.click(createContactButton);
-
-        await waitFor(() => {
-            expect(screen.getByText('Add Contact')).toBeVisible();
-        });
-    });
-
-    test('Phases widget is created when option is clicked', async () => {
-        useParamsMock.mockReturnValue({ engagementId: '1' });
-        getEngagementMock.mockReturnValueOnce(
-            Promise.resolve({
-                ...engagement,
-                surveys: surveys,
-            }),
-        );
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([]));
-        postWidgetMock.mockReturnValue(Promise.resolve(phasesWidget));
-        mockWidgetsRtkUnwrap.mockReturnValueOnce(Promise.resolve([phasesWidget]));
-        const { container } = render(<EngagementForm />);
-
-        await waitFor(() => {
-            expect(screen.getByDisplayValue('Test Engagement')).toBeInTheDocument();
-            expect(container.querySelector('span.MuiSkeleton-root')).toBeNull();
-        });
-
-        const addWidgetButton = screen.getByText('Add Widget');
-        fireEvent.click(addWidgetButton);
-
-        await waitFor(() => {
-            expect(screen.getByText('Select Widget')).toBeVisible();
-        });
-
-        const phasesOption = screen.getByTestId(`widget-drawer-option/${WidgetType.Phases}`);
-        fireEvent.click(phasesOption);
-
-        await waitFor(() => {
-            expect(screen.getByTestId('engagementPhaseSelect')).toBeVisible();
-        });
-
-        expect(postWidgetMock).toHaveBeenNthCalledWith(1, engagement.id, {
-            widget_type_id: WidgetType.Phases,
-            engagement_id: engagement.id,
-        });
-        expect(mockWidgetsRtkUnwrap).toHaveBeenCalledTimes(2);
-
-        const saveWidgetButton = screen.getByTestId('savePhasesWidgetButton');
-        expect(saveWidgetButton).toBeDisabled();
-
-        const standalonePhaseCheckbox = screen.getByTestId('standalonePhaseCheckbox');
-        fireEvent.click(standalonePhaseCheckbox);
-
-        fireEvent.click(saveWidgetButton);
-
-        await waitFor(() => {
-            expect(saveWidgetButton).not.toBeVisible();
-        });
-
-        expect(postWidgetItemMock).toHaveBeenNthCalledWith(1, phasesWidget.id, {
-            widget_id: phasesWidget.id,
-            widget_data_id: 0,
         });
     });
 });
