+++ met-web/tests/unit/components/widgets/DocumentWidget.test.tsx
@@ -1,8 +1,8 @@
 import { render, waitFor, screen, fireEvent } from '@testing-library/react';
 import React from 'react';
 import '@testing-library/jest-dom';
-import EngagementForm from '../../../src/components/engagement/form';
-import { setupEnv } from './setEnvVars';
+import EngagementForm from '../../../../src/components/engagement/form';
+import { setupEnv } from '../setEnvVars';
 import * as reactRedux from 'react-redux';
 import * as reactRouter from 'react-router';
 import * as engagementService from 'services/engagementService';
